缩放值10000

+++ Assets/Scripts/Msg/InGame/MsgSyncTank.cs
@@ -9,19 +9,19 @@ public class MsgSyncTank : MsgBase
     }
 
     //位置
-    public float x = 0;
-    public float y = 0;
-    public float z = 0;
+    public int x = 0;
+    public int y = 0;
+    public int z = 0;
 
     //旋转
-    public float ex = 0;
-    public float ey = 0;
-    public float ez = 0;
+    public int ex = 0;
+    public int ey = 0;
+    public int ez = 0;
 
     /// <summary>
     /// 炮塔旋转.y
     /// </summary>
-    public float turretY = 0;
+    public int turretY = 0;
 
     /// <summary>
     /// 服务端补充，哪个坦克
